[StateMachine, Notifications] fixes for Notifications in Admin Backend
