+++ src/CoreShop/Bundle/CoreBundle/Resources/install/data/notification-rules.json
@@ -83,9 +83,9 @@
         }
       },
       {
-        "type": "order.paymentState",
+        "type": "order.orderPaymentState",
         "configuration": {
-          "paymentState": "paid"
+          "orderPaymentState": "paid"
         }
       }
     ],
