+++ src/CoreShop/Bundle/CoreBundle/Resources/public/pimcore/js/notification/conditions/order/shipmentState.js
@@ -21,8 +21,8 @@ coreshop.notification.rule.conditions.shipmentState = Class.create(coreshop.rule
                 {
                     xtype: 'combo',
                     fieldLabel: t('coreshop_transition_direction_state'),
-                    name: 'shipmentState',
-                    value: this.data ? this.data.states : [],
+                    name: 'orderState',
+                    value: this.data ? this.data.orderState : [],
                     width: 250,
                     store: pimcore.globalmanager.get('coreshop_states_order_shipment'),
                     triggerAction: 'all',
