+++ src/CoreShop/Bundle/NotificationBundle/Resources/translations/admin.de.yml
@@ -1,35 +1,3 @@
-coreshop_carrier: 'Versanddienst'
-coreshop_paymentProvider: 'Zahlungsanbieter'
 coreshop_email_document: 'Email Dokument'
 coreshop_notification_rule: 'Notification Rule'
 coreshop_notification_rule_type: 'Regel-Typ'
-coreshop_condition_invoiceState: 'Rechnungsstatus'
-coreshop_condition_shipmentState: 'Versandstatus'
-coreshop_transition_direction_state: 'Zustand'
-coreshop_transition_direction_mode: 'Modus'
-coreshop_transition_to: 'Status ändert zu'
-coreshop_transition_from: 'Status ändert von'
-coreshop_transition_all: 'Alle Änderungen'
-coreshop_invoice_partial: 'Teilrechnung exsistiert'
-coreshop_invoice_full: 'Alle Rechnungen wurden erstellt'
-coreshop_invoice_all: 'Beliebiger Status trifft zu'
-coreshop_shipment_partial: 'Lieferschein für Teil-Versand wurde erstellt'
-coreshop_shipment_full: 'Alle Lieferscheine wurden erstellt'
-coreshop_shipment_all: 'Beliebiger Status trifft zu'
-coreshop_condition_userType: 'Benutzer Typ'
-coreshop_user_type_new: 'Registrierung'
-coreshop_user_type_password: 'Passwort zurücksetzen'
-coreshop_condition_messageType: 'Nachrichten Typ'
-coreshop_message_type_customer: 'Kundennachricht'
-coreshop_message_type_customer_reply: 'Kundennachricht Antwort'
-coreshop_message_type_contact: 'Kontakt'
-coreshop_grid_reorder: 'Drag and drop um die Sortierung zu ändern'
-coreshop_condition_paymentState: 'Zahlungsstatus'
-coreshop_payment_partial: 'Teilzahung gemacht'
-coreshop_payment_full: 'Gesamtzahlung gemacht'
-coreshop_permission_notification: 'Notification'
-coreshop_condition_orderState: 'Bestellstatus'
-coreshop_action_mail: 'Mail'
-coreshop_action_orderMail: 'Bestellung Mail'
-coreshop_mail_rule_send_invoices: 'Rechnung senden'
-coreshop_mail_rule_send_shipments: 'Lieferschein senden'
