+++ src/CoreShop/Bundle/NotificationBundle/Resources/translations/admin.en.yml
@@ -1,35 +1,3 @@
-coreshop_carrier: 'Carrier'
-coreshop_paymentProvider: 'Paymentprovider'
 coreshop_email_document: 'Email Document'
 coreshop_notification_rule: 'Notification Rule'
 coreshop_notification_rule_type: 'Rule Type'
-coreshop_condition_invoiceState: 'Invoice State'
-coreshop_condition_shipmentState: 'Shipment State'
-coreshop_transition_direction_state: 'State'
-coreshop_transition_direction_mode: 'Mode'
-coreshop_transition_to: 'State changes to'
-coreshop_transition_from: 'State changes from'
-coreshop_transition_all: 'All Changes'
-coreshop_invoice_partial: 'Partial'
-coreshop_invoice_full: 'Full'
-coreshop_invoice_all: 'All'
-coreshop_shipment_partial: 'Partial'
-coreshop_shipment_full: 'Full'
-coreshop_shipment_all: 'All'
-coreshop_condition_userType: 'User Type'
-coreshop_user_type_new: 'New Account'
-coreshop_user_type_password: 'Password Reset'
-coreshop_condition_messageType: 'Message Type'
-coreshop_message_type_customer: 'Customer'
-coreshop_message_type_customer_reply: 'Customer Reply'
-coreshop_message_type_contact: 'Contact'
-coreshop_grid_reorder: 'Drag and drop to reorganize'
-coreshop_condition_paymentState: 'Payment State'
-coreshop_payment_partial: 'Partial Payment done'
-coreshop_payment_full: 'Full Payment done'
-coreshop_permission_notification: 'Notification'
-coreshop_condition_orderState: 'Order State'
-coreshop_action_mail: 'Mail'
-coreshop_action_orderMail: 'Order Mail'
-coreshop_mail_rule_send_invoices: 'Send Invoices'
-coreshop_mail_rule_send_shipments: 'Send Shipments'
