+++ src/CoreShop/Bundle/OrderBundle/StateResolver/OrderDocumentStateResolver.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\StateResolver;
 
-use CoreShop\Component\Core\OrderPaymentStates;
+use CoreShop\Component\Order\OrderPaymentStates;
 use CoreShop\Component\Order\Model\OrderInterface;
 use CoreShop\Component\Order\OrderStates;
 use CoreShop\Component\Order\Processable\ProcessableInterface;
