+++ src/CoreShop/Bundle/OrderBundle/StateResolver/OrderPaymentStateResolver.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\StateResolver;
 
-use CoreShop\Component\Core\OrderPaymentTransitions;
+use CoreShop\Component\Order\OrderPaymentTransitions;
 use CoreShop\Component\Order\Model\OrderInterface;
 use CoreShop\Component\Order\StateResolver\StateResolverInterface;
 use CoreShop\Component\Payment\Model\PaymentInterface;
