+++ src/CoreShop/Bundle/OrderBundle/Templating/Helper/OrderStateHelper.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\Templating\Helper;
 
-use CoreShop\Component\Core\OrderPaymentStates;
+use CoreShop\Component\Order\OrderPaymentStates;
 use CoreShop\Component\Order\Model\OrderInterface;
 use CoreShop\Component\Order\OrderShipmentStates;
 use CoreShop\Component\Order\OrderStates;
