+++ src/CoreShop/Component/Core/Inventory/Operator/OrderInventoryOperator.php
@@ -13,7 +13,7 @@
 namespace CoreShop\Component\Core\Inventory\Operator;
 
 use CoreShop\Component\Core\Model\OrderInterface;
-use CoreShop\Component\Core\OrderPaymentStates;
+use CoreShop\Component\Order\OrderPaymentStates;
 use CoreShop\Component\Inventory\Model\StockableInterface;
 use CoreShop\Component\Order\Model\OrderItemInterface;
 use Doctrine\Common\Persistence\ObjectManager;
