+++ src/CoreShop/Component/Core/Order/Transformer/CartToSaleTransformer.php
@@ -6,7 +6,7 @@ use CoreShop\Component\Core\Model\CarrierInterface;
 use CoreShop\Component\Core\Model\CartInterface;
 use CoreShop\Component\Core\Model\OrderInterface;
 use CoreShop\Component\Core\Model\QuoteInterface;
-use CoreShop\Component\Core\OrderPaymentStates;
+use CoreShop\Component\Order\OrderPaymentStates;
 use CoreShop\Component\Currency\Converter\CurrencyConverterInterface;
 use CoreShop\Component\Order\Model\ProposalInterface;
 use CoreShop\Component\Order\Model\SaleInterface;
